+++ package.js
@@ -3,7 +3,7 @@
 Package.describe({
   name: 'quave:alert-react-tailwind',
   summary: 'Alert for React & Tailwind apps',
-  version: '1.0.1',
+  version: '1.0.2',
   git: 'https://github.com/quavedev/alert-react-tailwind',
 });
 
@@ -11,6 +11,7 @@ Package.onUse(api => {
   api.versionsFrom('2.5.3');
 
   api.use('ecmascript');
+  api.use('quave:settings@1.0.0');
 
   api.mainModule('alert-react-tailwind.js', 'client');
 });