+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 1.0.2 (2022-02-17)
+
+- Implements `autoCloseTimeout`.
+  - You can specify in the Meteor settings, like:
+  ```json
+  {
+    "packages": { 
+      "quave:alert-react-tailwind": {
+        "autoCloseTimeout": 5000
+      }
+    }
+  }
+    ```
+
 ## 1.0.1 (2022-01-19)
 
 - Fixes the default title in the popover that was in Portuguese instead of English.